Split up service tests into their own files
