+++ Rakefile
@@ -1,7 +1,7 @@
 require 'rake/testtask'
 
 Rake::TestTask.new do |t|
-  t.libs << 'test'
+  t.pattern = 'test/*_test.rb'
 end
 
 desc 'Run tests'
